SENA-109 Code Verification and Correction in the Project - Project

+++ LaCiguenia-Backend/src/main/java/LaCiguenia/commons/domains/wrapper/DetailProductForInvoice.java
@@ -10,7 +10,7 @@ public interface DetailProductForInvoice {
     String getProductName();
     Double getProductPrice();
     Integer getDetailAmount();
-    Integer getProductIva();
+    Double getProductIva();
     Double getDetailSubtotal();
     Double getInvoiceIva();
     Double getInvoiceTotal();
